requirements.txt
streamlit
pillow
google-genai
python-dotenv
chromadb
google-generativeai
pymupdf